+++ apps/hubble/src/storage/db/types.ts
@@ -53,12 +53,14 @@ export enum RootPrefix {
   HubEvents = 15,
   /* The network ID that the rocksDB was created with */
   Network = 16,
+  /* Used to store name proofs */
+  UserNameProof = 17,
   /* Used to store rent registry events */
-  RentRegistryEvent = 17,
+  RentRegistryEvent = 18,
   /* Index rent registry events by expiry */
-  RentRegistryEventsByExpiry = 18,
+  RentRegistryEventsByExpiry = 19,
   /* Used to store storage admin registry events */
-  StorageAdminRegistryEvent = 17,
+  StorageAdminRegistryEvent = 20,
 }
 
 /**