+++ apps/hubble/src/utils/logger.ts
@@ -8,6 +8,7 @@ import {
   RentRegistryEvent,
   StorageAdminRegistryEvent,
   storageRegistryEventTypeToJSON,
+  UserNameProof,
 } from '@farcaster/hub-nodejs';
 import pino from 'pino';
 
@@ -49,6 +50,7 @@ if (process.env['NODE_ENV'] === 'test' || process.env['CI']) {
 }
 
 export const logger = pino.pino(defaultOptions);
+export type Logger = pino.Logger;
 
 export const messageTypeToName = (type?: MessageType) => {
   if (!type) return '';
@@ -94,5 +96,13 @@ export const storageAdminRegistryEventToLog = (event: StorageAdminRegistryEvent)
     blockNumber: event.blockNumber,
     from: bytesToHexString(event.from)._unsafeUnwrap(),
     type: storageRegistryEventTypeToJSON(event.type),
+  };
+};
+
+export const usernameProofToLog = (usernameProof: UserNameProof) => {
+  return {
+    timestamp: usernameProof.timestamp,
+    name: Buffer.from(usernameProof.name).toString('utf-8').replace(/\0/g, ''),
+    owner: bytesToHexString(usernameProof.owner)._unsafeUnwrap(),
   };
 };