+++ packages/core/src/crypto/eip712.ts
@@ -1,5 +1,4 @@
-import type { Signer } from 'ethers';
-import { recoverAddress, TypedDataEncoder } from 'ethers';
+import { recoverAddress, TypedDataEncoder, Signer, TypedDataDomain, TypedDataField } from 'ethers';
 import { err, Result, ResultAsync } from 'neverthrow';
 import { bytesToHexString, hexStringToBytes } from '../bytes';
 import { HubAsyncResult, HubError, HubResult } from '../errors';
@@ -40,6 +39,19 @@ export const EIP_712_FARCASTER_MESSAGE_DATA = [
   },
 ];
 
+export const EIP_712_USERNAME_DOMAIN = {
+  name: 'Farcaster name verification',
+  version: '1',
+  chainId: 1,
+  verifyingContract: '0xe3be01d99baa8db9905b33a3ca391238234b79d1', // name registry contract, will be the farcaster ENS CCIP contract later
+};
+
+export const EIP_712_USERNAME_PROOF = [
+  { name: 'name', type: 'string' },
+  { name: 'timestamp', type: 'uint256' },
+  { name: 'owner', type: 'address' },
+];
+
 export const getSignerKey = async (signer: MinimalEthersSigner): HubAsyncResult<Uint8Array> => {
   return ResultAsync.fromPromise(signer.getAddress(), (e) => new HubError('unknown', e as Error)).andThen(
     hexStringToBytes
@@ -63,27 +75,22 @@ export const verifyVerificationEthAddressClaimSignature = (
   claim: VerificationEthAddressClaim,
   signature: Uint8Array
 ): HubResult<Uint8Array> => {
-  const signatureHexResult = bytesToHexString(signature);
-  if (signatureHexResult.isErr()) {
-    return err(signatureHexResult.error);
-  }
+  return recoverSignature(
+    EIP_712_FARCASTER_DOMAIN,
+    { VerificationClaim: EIP_712_FARCASTER_VERIFICATION_CLAIM },
+    claim,
+    signature
+  );
+};
 
-  // Recover address from signature
-  const recoveredHexAddress = Result.fromThrowable(
-    () =>
-      recoverAddress(
-        TypedDataEncoder.hash(
-          EIP_712_FARCASTER_DOMAIN,
-          { VerificationClaim: EIP_712_FARCASTER_VERIFICATION_CLAIM },
-          claim
-        ),
-        signatureHexResult.value
-      ),
-    (e) => new HubError('bad_request.invalid_param', e as Error)
-  )();
+export type UserNameProofClaim = {
+  name: string;
+  timestamp: number;
+  owner: string; // hex address
+};
 
-  // Convert hex recovered address to bytes
-  return recoveredHexAddress.andThen((hex) => hexStringToBytes(hex));
+export const verifyUserNameProof = (nameProof: UserNameProofClaim, signature: Uint8Array): HubResult<Uint8Array> => {
+  return recoverSignature(EIP_712_USERNAME_DOMAIN, { UserNameProof: EIP_712_USERNAME_PROOF }, nameProof, signature);
 };
 
 export const signMessageHash = async (hash: Uint8Array, signer: MinimalEthersSigner): HubAsyncResult<Uint8Array> => {
@@ -97,19 +104,28 @@ export const signMessageHash = async (hash: Uint8Array, signer: MinimalEthersSig
 };
 
 export const verifyMessageHashSignature = (hash: Uint8Array, signature: Uint8Array): HubResult<Uint8Array> => {
-  const signatureHexResult = bytesToHexString(signature);
+  return recoverSignature(
+    EIP_712_FARCASTER_DOMAIN,
+    { MessageData: EIP_712_FARCASTER_MESSAGE_DATA },
+    { hash },
+    signature
+  );
+};
 
+const recoverSignature = (
+  domain: TypedDataDomain,
+  types: Record<string, Array<TypedDataField>>,
+  claim: Record<string, any>,
+  signature: Uint8Array
+): HubResult<Uint8Array> => {
+  const signatureHexResult = bytesToHexString(signature);
   if (signatureHexResult.isErr()) {
     return err(signatureHexResult.error);
   }
 
   // Recover address from signature
   const recoveredHexAddress = Result.fromThrowable(
-    () =>
-      recoverAddress(
-        TypedDataEncoder.hash(EIP_712_FARCASTER_DOMAIN, { MessageData: EIP_712_FARCASTER_MESSAGE_DATA }, { hash }),
-        signatureHexResult.value
-      ),
+    () => recoverAddress(TypedDataEncoder.hash(domain, types, claim), signatureHexResult.value),
     (e) => new HubError('bad_request.invalid_param', e as Error)
   )();